+++ vnet/lib/vnet/core/interface_manager.rb
@@ -150,9 +150,6 @@ module Vnet::Core
       @dp_info.tunnel_manager.publish(ACTIVATE_INTERFACE, activate_params)
       @dp_info.filter2_manager.publish(ACTIVATE_INTERFACE, activate_params)
       @dp_info.interface_segment_manager.publish(ACTIVATE_INTERFACE, activate_params)
-
-      item.ingress_filtering_enabled &&
-        @dp_info.filter_manager.async.apply_filters(item_map)
     end
 
     def item_post_uninstall(item)
@@ -164,13 +161,6 @@ module Vnet::Core
       @dp_info.tunnel_manager.publish(DEACTIVATE_INTERFACE, deactivate_params)
       @dp_info.filter2_manager.publish(DEACTIVATE_INTERFACE, deactivate_params)
       @dp_info.interface_segment_manager.publish(DEACTIVATE_INTERFACE, deactivate_params)
-
-      @dp_info.filter_manager.async.remove_filters(item.id)
-
-      item.mac_addresses.each { |id, mac|
-        @dp_info.connection_manager.async.remove_catch_new_egress(id)
-        @dp_info.connection_manager.async.close_connections(id)
-      }
 
       deactivate_local_interface(item)
     end
@@ -274,9 +264,6 @@ module Vnet::Core
                            mac_address: mac_address,
                            segment_id: mac_lease.segment_id,
                            cookie_id: mac_lease.cookie_id)
-
-      item.ingress_filtering_enabled &&
-        @dp_info.connection_manager.async.catch_new_egress(item.id, mac_address)
     end
 
     # INTERFACE_RELEASED_MAC_ADDRESS on queue 'item.id'
@@ -288,9 +275,6 @@ module Vnet::Core
       return if mac_lease && mac_lease.interface_id == item.id
 
       item.remove_mac_address(mac_lease_id: params[:mac_lease_id])
-
-      @dp_info.connection_manager.async.remove_catch_new_egress(params[:mac_lease_id])
-      @dp_info.connection_manager.async.close_connections(params[:mac_lease_id])
     end
 
     # INTERFACE_LEASED_IPV4_ADDRESS on queue 'item.id'