+++ vnet/lib/vnet/openflow/controller.rb
@@ -67,6 +67,8 @@ module Vnet::Openflow
       return unless dp_info
 
       case message.cookie >> COOKIE_PREFIX_SHIFT
+      when COOKIE_PREFIX_FILTER2
+        dp_info.filter2_manager.async.packet_in(message)
       when COOKIE_PREFIX_INTERFACE
         dp_info.interface_manager.async.packet_in(message)
       when COOKIE_PREFIX_TRANSLATION
@@ -77,8 +79,6 @@ module Vnet::Openflow
         dp_info.segment_manager.async.packet_in(message)
       when COOKIE_PREFIX_SERVICE
         dp_info.service_manager.async.packet_in(message)
-      when COOKIE_PREFIX_CONNECTION
-        dp_info.connection_manager.async.packet_in(message)
       end
     end